requirements.txt
asgiref==3.8.1
boto3==1.35.82
botocore==1.35.82
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-database-url==2.3.0
dj-email-url==1.0.6
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-appconf==1.0.6
django-cache-url==3.4.5
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-filter==24.3
django-imagekit==5.0.0
django-phonenumber-field==8.0.0
django-storages==1.14.4
djangorestframework==3.15.2
environs==11.1.0
idna==3.10
jmespath==1.0.1
marshmallow==3.23.1
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.49
pilkit==3.0
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
